+++ packages/components/Menu/src/MenuItemCheckbox/MenuItemCheckbox.types.ts
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { ColorValue } from 'react-native';
 import { XmlProps } from 'react-native-svg';
 import type { IViewProps } from '@fluentui-react-native/adapters';
-import { TextProps } from '@fluentui-react-native/experimental-text';
+import { TextProps } from '@fluentui-react-native/text';
 import { IPressableHooks } from '@fluentui-react-native/interactive-hooks';
 import { MenuItemProps, MenuItemTokens } from '../MenuItem/MenuItem.types';