+++ packages/components/Notification/src/Notification.types.ts
@@ -2,16 +2,17 @@ import { PressableProps } from 'react-native';
 import { IconProps, IconSourcesType } from '@fluentui-react-native/icon';
 import { IViewProps, ITextProps } from '@fluentui-react-native/adapters';
 import { ButtonProps } from '@fluentui-react-native/button';
-import { SvgProps } from 'react-native-svg';
 import { FontTokens, IBorderTokens, IColorTokens, LayoutTokens } from '@fluentui-react-native/tokens';
 import { InteractionEvent } from '@fluentui-react-native/interactive-hooks';
-import { ActionButtonColorStates } from './Notification.helper';
+import { NotificationButtonColorStates } from './Notification.helper';
+import { ShadowProps } from '@fluentui-react-native/experimental-shadow';
+import { ShadowToken } from '@fluentui-react-native/theme-types';
 
 export const notification = 'Notification';
 export const NotificationVariants = ['primary', 'neutral', 'primaryBar', 'primaryOutlineBar', 'neutralBar', 'danger', 'warning'] as const;
 export type NotificationVariant = typeof NotificationVariants[number];
 
-export interface NotificationTokens extends LayoutTokens, IBorderTokens, IColorTokens, FontTokens, ActionButtonColorStates {
+export interface NotificationTokens extends LayoutTokens, IBorderTokens, IColorTokens, FontTokens, NotificationButtonColorStates {
   primary: NotificationTokens;
   neutral: NotificationTokens;
   primaryBar: NotificationTokens;
@@ -21,6 +22,7 @@ export interface NotificationTokens extends LayoutTokens, IBorderTokens, IColorT
   warning: NotificationTokens;
   hasTitle: NotificationTokens;
   isBar: NotificationTokens;
+  shadowToken?: ShadowToken;
 }
 
 export interface NotificationProps {
@@ -28,10 +30,31 @@ export interface NotificationProps {
    * Notification variants: 'primary' | 'neutral' | 'primaryBar' | 'primaryOutlineBar' | 'neutralBar' | 'danger' | 'warning'
    */
   variant: NotificationVariant;
+
+  /**
+   * Optional icon that appears on the left side of toast notifications
+   */
   icon?: IconSourcesType;
+
+  /**
+   * Optional icon that appears at the top of toast notifications
+   */
   title?: string;
+
+  /**
+   * Text in the action button
+   */
   action?: string;
+
+  /**
+   * Callback function that is triggered by pressing the entire notification
+   */
   onPress?: (e: InteractionEvent) => void;
+
+  /**
+   * Callback function that is triggered by pressing the action button
+   */
+  onActionPress?: (e: InteractionEvent) => void;
 }
 
 export interface NotificationSlotProps {
@@ -41,7 +64,7 @@ export interface NotificationSlotProps {
   title?: ITextProps;
   message: ITextProps;
   action?: ButtonProps;
-  dismissIcon?: SvgProps;
+  shadow?: ShadowProps;
 }
 
 export interface NotificationType {