+++ packages/components/Notification/src/NotificationTokens.ts
@@ -2,6 +2,16 @@ import { Theme } from '@fluentui-react-native/framework';
 import { TokenSettings } from '@fluentui-react-native/use-styling';
 import { NotificationTokens } from './Notification.types';
 
+const emptyShadowStyle = {
+  ambient: { x: 0, y: 0, blur: 0, color: '#00000000' },
+  key: { x: 0, y: 0, blur: 0, color: '#00000000' },
+};
+
+const notificationShadowStyle = {
+  ambient: { x: 0, y: 8, blur: 8, color: '#00000024' },
+  key: { x: 0, y: 0, blur: 1, color: '#0000001f' },
+};
+
 export const defaultNotificationTokens: TokenSettings<NotificationTokens, Theme> = (t: Theme) =>
   ({
     backgroundColor: t.colors.background,
@@ -10,13 +20,13 @@ export const defaultNotificationTokens: TokenSettings<NotificationTokens, Theme>
     borderWidth: 1,
     color: t.colors.brandForeground1,
     fontFamily: 'primary',
-    fontLetterSpacing: -0.24, // iOS only prop
     fontLineHeight: 20,
     fontSize: 15,
     fontWeight: '600',
+    minHeight: 64,
     padding: 16,
+    shadowToken: notificationShadowStyle,
     hasTitle: {
-      fontLetterSpacing: -0.08, // iOS only prop
       fontLineHeight: 18,
       fontSize: 13,
       fontWeight: '400',
@@ -25,54 +35,53 @@ export const defaultNotificationTokens: TokenSettings<NotificationTokens, Theme>
     isBar: {
       borderRadius: 0,
       fontWeight: '400',
+      shadowToken: emptyShadowStyle,
     },
     primary: {
       backgroundColor: '#EBF3FC', // brandBackground4
       color: '#0F6CBD', // brandForeground4
       disabledColor: '#2886DE', // brandForegroundDisabled
       /**
-       * None of the foreground tokens here have pressed versions so the foreground color with an alpha value was used.
+       * None of the foreground tokens here have pressed versions so the foreground color with an alpha value is used.
        * The FluentUI Apple NotificationView was used to color match.
        */
-      pressedColor: '#0F6CBD30',
+      pressedColor: '#0F6CBD30', // opacity: 0.19
     },
     neutral: {
       backgroundColor: '#FAFAFA', // background4
       color: '#616161', // foreground2
       disabledColor: '#FFFFFF', // foregroundDisabled2
-      pressedColor: '#61616145',
-    },
-    // hardcoded values from FluentUI Apple NotificationView
-    danger: {
-      backgroundColor: '#FDF6F6',
-      color: '#BC2F34',
-      pressedColor: '#BC2F3433',
-    },
-    // hardcoded values from FluentUI Apple NotificationView
-    warning: {
-      backgroundColor: '#FFFBD6',
-      color: '#4C4400',
-      pressedColor: '#4C440033',
+      pressedColor: '#61616145', // opacity: 0.27
     },
     primaryBar: {
       backgroundColor: '#EBF3FC', // brandBackground4
       borderWidth: 0,
       color: '#0F6CBD', // brandForeground4
       disabledColor: '#2886DE', // brandForegroundDisabled
-      pressedColor: '#0F6CBD30',
+      pressedColor: '#0F6CBD30', // opacity: 0.19
     },
     primaryOutlineBar: {
       backgroundColor: '#FFFFFF', // background1
       borderColor: '#E0E0E0', // stroke2
       color: '#0F6CBD', // brandForeground1
       disabledColor: '#2886DE', // brandForegroundDisabled
-      pressedColor: '#0F6CBD30',
+      pressedColor: '#0F6CBD30', // opacity: 0.19
     },
     neutralBar: {
       backgroundColor: '#F0F0F0', // background5
       borderWidth: 0,
       color: '#616161', // foreground2
       disabledColor: '#FFFFFF', // foregroundDisabled2
-      pressedColor: '#61616145',
+      pressedColor: '#61616145', // opacity: 0.27
+    },
+    danger: {
+      backgroundColor: '#FDF6F6', // PaletteRedBackground1
+      color: '#9F282C', // PaletteRedForeground1
+      pressedColor: '#9F282C32', // opacity: 0.2
+    },
+    warning: {
+      backgroundColor: '#FFFBD6', // PaletteYellowBackground1
+      color: '#4C4400', // PaletteYellowForeground1
+      pressedColor: '#4C440033', // opacity: 0.2
     },
   } as NotificationTokens);