+++ packages/experimental/Badge/src/Badge.styling.ts
@@ -1,31 +1,21 @@
-import { badgeName, BadgeCoreTokens, BadgeTokens, BadgeSlotProps, BadgeProps } from './Badge.types';
+import {
+  badgeName,
+  BadgeCoreTokens,
+  BadgeTokens,
+  BadgeSlotProps,
+  BadgeProps,
+  BadgeColors,
+  BadgeSizes,
+  BadgeShapes,
+  BadgeAppearances,
+} from './Badge.types';
 import { UseStylingOptions, buildProps, Theme } from '@fluentui-react-native/framework';
-import { borderStyles, layoutStyles } from '@fluentui-react-native/tokens';
+import { borderStyles, layoutStyles, fontStyles } from '@fluentui-react-native/tokens';
 import { defaultBadgeTokens } from './BadgeTokens';
 import { defaultBadgeColorTokens } from './BadgeColorTokens';
 
-export const coreBadgeStates: (keyof BadgeCoreTokens)[] = [
-  'smallest',
-  'smaller',
-  'small',
-  'medium',
-  'large',
-  'largest',
-  'rounded',
-  'circular',
-  'square',
-  'iconColor',
-];
-export const badgeStates: (keyof BadgeTokens)[] = [
-  'hovered',
-  'focused',
-  'filled',
-  'outline',
-  'tint',
-  'ghost',
-  'filledInverted',
-  ...coreBadgeStates,
-];
+export const coreBadgeStates: (keyof BadgeCoreTokens)[] = [...BadgeSizes, ...BadgeShapes];
+export const badgeStates: (keyof BadgeTokens)[] = [...coreBadgeStates, ...BadgeColors, ...BadgeAppearances];
 
 export const stylingSettings: UseStylingOptions<BadgeProps, BadgeSlotProps, BadgeTokens> = {
   tokens: [defaultBadgeTokens, defaultBadgeColorTokens, badgeName],
@@ -35,12 +25,11 @@ export const stylingSettings: UseStylingOptions<BadgeProps, BadgeSlotProps, Badg
       (tokens: BadgeTokens, theme: Theme) => ({
         style: {
           ...getBadgePosition(tokens),
-          display: 'flex',
           alignItems: 'center',
           flexDirection: 'row',
           alignSelf: 'flex-start',
           justifyContent: 'center',
-          height: tokens.height,
+          minHeight: tokens.minHeight,
           width: tokens.width,
           backgroundColor: tokens.backgroundColor,
           ...borderStyles.from(tokens, theme),
@@ -50,21 +39,20 @@ export const stylingSettings: UseStylingOptions<BadgeProps, BadgeSlotProps, Badg
       ['backgroundColor', 'width', 'height', 'bottom', 'right', 'top', 'left', ...borderStyles.keys, ...layoutStyles.keys],
     ),
     icon: buildProps(
-      (tokens: BadgeTokens, theme: Theme) => ({
-        style: {
-          height: tokens.iconSize,
-          width: tokens.iconSize,
-          ...layoutStyles.from(tokens, theme),
-        },
+      (tokens: BadgeTokens) => ({
+        color: tokens.iconColor || tokens.color,
+        height: tokens.iconSize,
+        width: tokens.iconSize,
       }),
-      ['width', 'height', ...layoutStyles.keys],
+      ['width', 'height', 'iconSize', 'iconColor', 'color'],
     ),
     text: buildProps(
-      (tokens: BadgeTokens) => ({
-        variant: tokens.variant,
+      (tokens: BadgeTokens, theme: Theme) => ({
+        ...fontStyles.from(tokens, theme),
         color: tokens.color,
+        paddingHorizontal: tokens.textPadding,
       }),
-      ['variant', 'color'],
+      ['color', 'textPadding', ...fontStyles.keys],
     ),
   },
 };