+++ packages/experimental/Checkbox/CHANGELOG.md
@@ -1,8 +1,73 @@
 # Change Log - @fluentui-react-native/experimental-checkbox
 
-This log was last generated on Thu, 14 Jul 2022 20:03:37 GMT and should not be manually modified.
+This log was last generated on Tue, 02 Aug 2022 20:08:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.9
+
+Tue, 02 Aug 2022 20:08:43 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.16.9
+
+## 0.12.8
+
+Fri, 29 Jul 2022 22:47:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.16.8
+
+## 0.12.7
+
+Fri, 29 Jul 2022 21:04:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.16.7
+
+## 0.12.6
+
+Thu, 28 Jul 2022 23:57:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.16.6
+
+## 0.12.5
+
+Tue, 26 Jul 2022 19:30:23 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.16.5
+
+## 0.12.4
+
+Mon, 25 Jul 2022 21:27:20 GMT
+
+### Patches
+
+- Combine platform specific test apps (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/checkbox to v0.16.4
+
+## 0.12.3
+
+Sat, 23 Jul 2022 00:28:48 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.16.3
+
+## 0.12.2
+
+Tue, 19 Jul 2022 18:10:02 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.16.2
 
 ## 0.12.1