+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Thu, 14 Jul 2022 18:09:51 GMT and should not be manually modified.
+This log was last generated on Fri, 29 Jul 2022 21:04:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.3
+
+Fri, 29 Jul 2022 21:04:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.13.0
+
+## 0.13.2
+
+Thu, 28 Jul 2022 23:57:03 GMT
+
+### Patches
+
+- Delete workaround (ruaraki@microsoft.com)
+
+## 0.13.1
+
+Tue, 26 Jul 2022 19:30:23 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.12.1
 
 ## 0.13.0